+++ dataland-backend/src/main/kotlin/org/dataland/datalandbackend/controller/EuTaxonomyDataController.kt
@@ -2,7 +2,7 @@ package org.dataland.datalandbackend.controller
 
 import com.fasterxml.jackson.databind.ObjectMapper
 import org.dataland.datalandbackend.interfaces.DataStoreInterface
-import org.dataland.datalandbackend.model.EuTaxonomyDataSet
+import org.dataland.datalandbackend.model.EuTaxonomyData
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.web.bind.annotation.RequestMapping
 import org.springframework.web.bind.annotation.RestController
@@ -17,4 +17,4 @@ import org.springframework.web.bind.annotation.RestController
 class EuTaxonomyDataController(
     @Autowired var myDataStore: DataStoreInterface,
     @Autowired var myObjectMapper: ObjectMapper
-) : DataController<EuTaxonomyDataSet>(myDataStore, myObjectMapper, EuTaxonomyDataSet::class.java)
+) : DataController<EuTaxonomyData>(myDataStore, myObjectMapper, EuTaxonomyData::class.java)
